Fix rollback issue
The version of goleveldb that we were using before returned `nil`
instead of `[]byte{}`. Unfortunately, this changes the behavior in v0.20.0
of `github.com/cosmos/iavl`. The sdk and iavl interpret `nil` as
"doesn't exist", which meant that the iavl mistakenly couldn't rollback
to past states if the tree was empty (which happened for modules that
had an empty genesis state and no state updates thereafter).

We aren't adding a regression test because this only occurs in
goleveldb, which is technically hidden behind an interface.