Remove the cli in favor of a proper `node` package
This commit replaces the cli, which ran Monomer and the test app, with a
Node package that only encapsulates Monomer (and hides the test app behind
an interface).

It also simplifies the http server creation and tear-down.

Lastly, we write a simple test for the node package that ensures we can
set up the node and query the eth_chainId endpoint.