Improve shutdown process (#22)
The crux of the changes here is the addition of the `environment`
package, which wraps a single `conc.WaitGroup` and a slice of functions.
Conceptually, it is like a "main thread", which spawns worker threads and
executes cleanup operations when the application exits.

This removes the need for complicated `context.CancelCause`, `defer`, and
goroutine gymnastics typically required for proper cleanup. The last
place where `context.CancelCause` was used was in `node/http.go`, which
we remove with an error channel.

Most importantly, these changes allow us to add useful functionality
on the `node.Node` in the future. This was hazardous before, since we
couldn't know when services were started in `node.*Node.Start`.