Remove race condition in builder test (#26)
It was possible that the event bus subscription channel had not received
all of the events by the time we were requiring it. Removing the check
ensures we wait for the subscription channel to receive each event.

I'm not sure how we can ensure additional message aren't sent on the
subscription. That is left untested, as before.