Fix block hash bug
The OP stack requires blocks returned from the `eth` API to be internally consistent.
For example: the `block_hash` in the json must represent the hash of the
Ethereum block, as defined by Ethereum. Monomer was using a slightly different
block hash algorithm that excluded the tx root.

This changes Monomer's block hash algorithm to match Ethereum's.
Furthermore, we copy op-geth's rpc marshalling code to ensure
compatbility.

There are still errors in the e2e test (op-batcher not able to sign txs).
Those can be fixed in follow-ups.

Finally, we make a small fix in the adapters to make the tests pass.