Add contract for storing L2 cosmos app state in monomer EVM (#70)
Adds a solidity contract and corresponding go bindings to hold the current L2 cosmos app state in contract storage. This is necessary for implementing withdrawals because we need a valid account proof for the L2ToL1MessagePasser via the top-level Monomer state root. The valid account proof is needed for withdrawals and the top-level monomer state root is what the proposer stores in the L2OutputOracle contract on L1. Additionally, adds Makefile targets for easily instaling binding dependencies and for generating solidity bindings.