Add message nonce to the InitiateWithdrawal event attributes (#98)
In order to prove that a withdrawal occurred, a relayer will need to include the full
withdrawal tx information in their proof. Unfortunately, the message nonce is calculated
inside of the L2ToL1MessagePasser so it isn't available whenever the x/rollup event for
initiating the withdrawal is created.

To allow relayers to have the full withdrawal tx, the message nonce is appended onto the
tx's event attributes after the withdrawal hash is stored in the EVM. Additionally, some
other attributes like gasLimit and data were added as well to the InitiateWithdrawal
event in the x/rollup module make it easy for the relayer to build their withdrawal proof.