Update docs for latest Cosmos SDK version
We recently expunged Pulsar from our repo (yay!). In the process, we
needed to upgrade to the next patch version of the Cosmos SDK (v0.50.8).

Version v0.50.8 has dependencies on modules that rely on `gogoproto`'s
`MergedRegistry` function. Thus, it's not possible to comment out the
line (as we were doing [before]), unless we wanted to multiply the
number of `go replace` directives.

Instead, we address the root cause in the `go-libp2p` package. That's
the only new replace directive we need. We have [upstreamed the fix],
but that may not be used by the op-node for quite some time. A single
`replace` is fine for now.

This commit updates the docs with the correct instructions for the
latest Monomer changes.

[before]: https://github.com/polymerdao/monomer/blob/c7ecbee46914c943bfbecaf2bd5ded3076df09f2/doc/docs/learn/create-an-app-with-monomer.md?plain=1#L150
[upstreamed the fix]: https://github.com/libp2p/go-libp2p/pull/2920