Fix e2e test rollbacks
Derivation will fail [here] if the sequencer does not post with the
same address as the one specified in the onchain `SystemConfig`. Unfortunately,
this cannot be counted as a fatal error during derivation since anyone can
technically post batcher transactions. As a result, derivation will silently
continue with empty batches, which manifests itself in rollbacks every
time unsafe block consolidation is attempted when the sequencing window expires.

The `SystemConfig` is set when we convert the `DeployConfig` to a `RollupConfig`.
Thus, to fix the issue, we set the `BatcherSenderAddress` in the `DeployConfig`.

[here]: https://github.com/ethereum-optimism/optimism/blob/24a8d3e06e61c7a8938dfb7a591345a437036381/op-node/rollup/derive/data_source.go#L110-L114