Fix race condition in deposit e2e test
We are currently only waiting one L2 block for a deposit transaction
to be processed. However, it needs to first land on L1 before it is
derived and processed on L2. We weren't waiting long enough before.
This commit adds an additional `WaitL1` call before the `WaitL2` call
to ensure (with high probability) that the deposit transaction will
be processed on both L1 and L2 before we make assertions.