Add ParentHash to CometBFT header conversion
- Include ParentHash in LastBlockID field of CometBFT header
- Set PartSetHeader Total to 0 and Hash to nil

This change ensures that the parent block information is properly represented when converting from Monomer to CometBFT headers. Including the ParentHash in the LastBlockID allows for correct block linking in
the CometBFT chain.

The PartSetHeader is set with Total: 0 and Hash: nil  because our implementation doesn't use CometBFT's block partitioning feature.

https://github.com/cometbft/cometbft/blob/v0.37.x/spec/core/data_structures.md#header