Fix Comet /status RPC endpoint
Currently, the CometBFT /status RPC endpoint errors
out because no start block is provided. The head block
on startup is passed into the status API as the earliest
block to show up when hitting the /status endpoint.

In the future, a new DB method should be exposed to get the
earliest stored block instead of using the head block on
startup.