Enable transfers during L1 deposits
Updates the x/rollup deposit flow to transfer minted
ETH to the `To` address in the deposit tx. The amount
of transferred ETH is determined by the `Value` field
in the deposit tx.

Any extra minted ETH, determined by the `Mint` field,
will be sent to the sender address of the deposit tx.