Add InitiateFeeWithdrawal msg_server impl
Adds a message server implementation for the InitiateFeeWithdrawal
message to mimic the behavior of the OPStack's FeeVault contracts.

Users can permissionlessly triggers fee withdrawals on L2 if the
FeeCollector module account balance is above the predefined minimum
withdrawal amount. The fees are then able to be withdrawn to the
predefined target address on L1.